ai-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.1
pymongo==4.6.1
motor==3.3.2
google-generativeai==0.3.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
tensorflow==2.15.0
transformers==4.37.2
httpx==0.26.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiohttp==3.9.3
redis==5.0.1
loguru==0.7.2
pytest==8.0.0
pytest-asyncio==0.23.4
